Remove animation from page__hero--overlay image
The animation i.e. the intro delay of 0.3s on the `page__hero--overlay` looks as if the image is extremly huge and takes a while to load. During this time we have the white heading text on a white background. To fix this situation for cases in which the 284Kb image really takes a while to load, I changed the background-color of the heading to the blue used in the background-image.